requirements.txt
python-dotenv==1.0.1
firecrawl-py==0.0.12
playwright==1.44.0
boto3==1.34.3
Flask==3.0.0
openai==1.3.4
pandas==2.0.3
urlextract==1.8.0
requests~=2.31.0
asgiref==3.8.1
uuid==1.30